refactor: split backend sales logic from inventory app
