+++ backend/sales/apps.py
@@ -1,0 +1,10 @@
+from django.apps import AppConfig
+
+
+class SalesConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'sales'
+    verbose_name = 'Sales and Invoices'
+
+    def ready(self):
+        import sales.signals
